+++ src/main/java/hexlet/code/service/UserService.java
@@ -12,8 +12,10 @@ import org.springframework.stereotype.Service;
 import hexlet.code.dto.user.UserCreateDTO;
 import hexlet.code.dto.user.UserDTO;
 import hexlet.code.dto.user.UserUpdateDTO;
+import hexlet.code.exception.ResourceDeletionException;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.UserMapper;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.repository.UserRepository;
 import lombok.AllArgsConstructor;
 
@@ -28,9 +30,14 @@ public class UserService implements UserDetailsService {
     private UserMapper userMapper;
 
     @Autowired
+    private TaskRepository taskRepository;
+
+    @Autowired
     private PasswordEncoder passwordEncoder;
 
-    // GET /api/users
+    /**
+     * GET /api/users
+     */
     public List<UserDTO> index() {
         var users = userRepository.findAll();
         var ret = users.stream()
@@ -39,14 +46,18 @@ public class UserService implements UserDetailsService {
         return ret;
     }
 
-    // GET /api/users/{id}
+    /**
+     * GET /api/users/{id}
+     */
     public UserDTO show(long id) {
         var user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Not Found"));
         return userMapper.map(user);
     }
 
-    // POST /api/users
+    /**
+     * POST /api/users
+     */
     public UserDTO create(UserCreateDTO dto) {
         var newUser = userMapper.map(dto);
         newUser.setPasswordDigest(passwordEncoder.encode(dto.getPassword()));
@@ -55,7 +66,9 @@ public class UserService implements UserDetailsService {
         return userMapper.map(newUser);
     }
 
-    // PUT /api/users/{id}
+    /**
+     * PUT /api/users/{id}
+     */
     public UserDTO update(long id, UserUpdateDTO dto) {
         var user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Not Found"));
@@ -67,7 +80,13 @@ public class UserService implements UserDetailsService {
         return userMapper.map(user);
     }
 
+    /**
+     * DELETE /api/users/{id}
+     */
     public void delete(long id) {
+        if (taskRepository.existsByAssigneeId(id)) {
+            throw new ResourceDeletionException("Нельзя удалить пользователя, у него есть задача");
+        }
         userRepository.deleteById(id);
     }
